Using generators internally + refactoring flow
In preparation to add more async logic here (such as working with the cache)
I've decided to change how this module is structured internally. Primarilly,
it now uses generators instead of normal callbacks.

I'm using `unyield` so the API _technically_ hasn't changed, although it does
now require the `--harmony-generators` flag since it is using generators under
the hood. I've also removed `options.retries`, as that is an internal detail
that doesn't really need to be configurable at this time.